refactor benchmarks
